Complete Session 1: Formula Engine Core + Phase 6 Advanced Features + Phase 7 Quality & Tooling

+++ src/diagnostics.rs
@@ -9,18 +9,18 @@ pub fn format_error(source: &str, error: &FormulaError) -> String {
         // แยก source เป็นบรรทัด
         let lines: Vec<&str> = source.lines().collect();
 
-        if span.start.line > 0 && (span.start.line as usize) <= lines.len() {
-            let line_idx = span.start.line as usize - 1;
+        if span.start.line > 0 && span.start.line <= lines.len() {
+            let line_idx = span.start.line - 1;
             let line = lines[line_idx];
 
             // แสดงบรรทัดที่มี error
             output.push_str(&format!(" {:4} | {}\n", span.start.line, line));
 
             // แสดง caret ชี้ตำแหน่ง
-            let caret_col = span.start.column as usize;
+            let caret_col = span.start.column;
             if caret_col <= line.len() {
                 let spaces = " ".repeat(7 + caret_col.saturating_sub(1)); // 7 คือ "XXXX | "
-                let carets = "^".repeat((span.end.column - span.start.column + 1).max(1) as usize);
+                let carets = "^".repeat((span.end.column - span.start.column + 1).max(1));
                 output.push_str(&format!("{}{}\n", spaces, carets));
             }
         }
